fix: address PR review feedback for multi-draft apps
Filter draft listing to the current user, simplify draft ownership helpers,
move create/delete tests to their own files, and remove the redundant
SaveCanvasDraftInTransaction wrapper.
