+++ pkg/grpc/actions/canvases/update_canvas_version.go
@@ -153,7 +153,7 @@ func UpdateCanvasVersionWithUsage(
 
 		nodes := injectMetadataIntoNodes(version.Nodes, nodes)
 
-		if err := ensureVersionIsOwnedRegisteredDraftInTransaction(tx, canvasUUID, userUUID, version); err != nil {
+		if err := ensureVersionIsOwnedRegisteredDraft(userUUID, version); err != nil {
 			return err
 		}
 
@@ -204,12 +204,7 @@ func UpdateCanvasVersionWithUsage(
 	}, nil
 }
 
-func ensureVersionIsOwnedRegisteredDraftInTransaction(
-	_ *gorm.DB,
-	_ uuid.UUID,
-	userID uuid.UUID,
-	version *models.CanvasVersion,
-) error {
+func ensureVersionIsOwnedRegisteredDraft(userID uuid.UUID, version *models.CanvasVersion) error {
 	if version.OwnerID == nil || *version.OwnerID != userID {
 		return status.Error(codes.PermissionDenied, "version owner mismatch")
 	}
